修改保存

+++ mooc-core/src/main/java/cn/edu/lingnan/mooc/core/repository/LoginLogRepository.java
@@ -15,7 +15,7 @@ import java.util.List;
  * @see JpaRepository 支持常用增删查改
  * @see JpaSpecificationExecutor 支持多条件分页
  */
-public interface LoginLogRepository extends JpaRepository<LoginLog, Integer>,JpaSpecificationExecutor<LoginLog> {
+public interface LoginLogRepository extends JpaRepository<LoginLog, Long>,JpaSpecificationExecutor<LoginLog> {
 
 
     /**
